// src/app/common/types/ProfilePropertyFilters.ts
export default interface ProfilePropertyFilters {
  Address: string;
  City: string;
  Neighborhood: string;
  Status: string;
  Building: string;
  AddressSort: number;
  CitySort: number;
  NeighborhoodSort: number;
  SizeSort: number;
  CollectionSort: number;
  MintSort: number;
  BoostSort: number;
  StatusSort: number;
  BuildingSort: number;
  MintedSort: number;
  AcquiredOnSort: number;
  PageNumber: number;
}